+++ compiler-plugin/src/main/java/io/ballerina/stdlib/graphql/compiler/Utils.java
@@ -29,18 +29,13 @@ import io.ballerina.compiler.api.symbols.TypeSymbol;
 import io.ballerina.compiler.api.symbols.UnionTypeSymbol;
 import io.ballerina.compiler.syntax.tree.FunctionDefinitionNode;
 import io.ballerina.projects.plugins.SyntaxNodeAnalysisContext;
-import io.ballerina.tools.diagnostics.Diagnostic;
-import io.ballerina.tools.diagnostics.DiagnosticFactory;
-import io.ballerina.tools.diagnostics.DiagnosticInfo;
-import io.ballerina.tools.diagnostics.DiagnosticSeverity;
-import io.ballerina.tools.diagnostics.Location;
 
 import java.util.Optional;
 
 /**
  * Util class for the compiler plugin.
  */
-public class Utils {
+public final class Utils {
 
     private Utils() {
     }
@@ -50,71 +45,9 @@ public class Utils {
     public static final String PACKAGE_ORG = "ballerina";
 
     // resource function constants
-    public static final String RESOURCE_FUNCTION_GET = "get";
     public static final String LISTENER_IDENTIFIER = "Listener";
-    public static final String DOUBLE_UNDERSCORES = "__";
     public static final String CONTEXT_IDENTIFIER = "Context";
-
-    /**
-     * Compilation errors.
-     */
-    enum CompilationError {
-        INVALID_FUNCTION("Remote methods are not allowed inside the service classes returned from GraphQL resources",
-                         "GRAPHQL_101", DiagnosticSeverity.ERROR),
-        INVALID_RETURN_TYPE("Invalid GraphQL field Type", "GRAPHQL_102", DiagnosticSeverity.ERROR),
-        INVALID_RESOURCE_INPUT_PARAM("Invalid input parameter type for GraphQL resource/remote function", "GRAPHQL_103",
-                                     DiagnosticSeverity.ERROR),
-        INVALID_RETURN_TYPE_NIL("A GraphQL field must have a return type", "GRAPHQL_104", DiagnosticSeverity.ERROR),
-        INVALID_RETURN_TYPE_ERROR_OR_NIL("A GraphQL field must have a return data type", "GRAPHQL_105",
-                                         DiagnosticSeverity.ERROR),
-        INVALID_RESOURCE_FUNCTION_ACCESSOR(
-                "Only \"" + RESOURCE_FUNCTION_GET + "\" accessor is allowed for GraphQL resource function",
-                "GRAPHQL_106", DiagnosticSeverity.ERROR),
-        INVALID_MULTIPLE_LISTENERS("A GraphQL service cannot be attached to multiple listeners", "GRAPHQL_107",
-                                   DiagnosticSeverity.ERROR),
-        INVALID_RETURN_TYPE_ERROR("A GraphQL field must have a return data type", "GRAPHQL_108",
-                                  DiagnosticSeverity.ERROR),
-        INVALID_LISTENER_INIT("http:Listener and graphql:ListenerConfiguration are mutually exclusive", "GRAPHQL_109",
-                              DiagnosticSeverity.ERROR),
-        INVALID_RETURN_TYPE_MULTIPLE_SERVICES("GraphQL union type cannot consist non-distinct services", "GRAPHQL_110",
-                                              DiagnosticSeverity.ERROR),
-        INVALID_FIELD_NAME("A GraphQL field name must not begin with \"" + DOUBLE_UNDERSCORES +
-                                   "\", which is reserved by GraphQL introspection", "GRAPHQL_111",
-                           DiagnosticSeverity.ERROR),
-        INVALID_RETURN_TYPE_ANY(
-                "A GraphQL field cannot have \"any\" or \"anydata\" as the type, instead use specific types",
-                "GRAPHQL_112", DiagnosticSeverity.ERROR),
-        MISSING_RESOURCE_FUNCTIONS("A GraphQL service must have at least one resource function", "GRAPHQL_113",
-                                   DiagnosticSeverity.ERROR),
-        INVALID_RETURN_TYPE_INPUT_OBJECT("A GraphQL field type cannot be an input type", "GRAPHQL_113",
-                                         DiagnosticSeverity.ERROR),
-        INVALID_RESOURCE_INPUT_OBJECT_PARAM("A GraphQL field cannot use an output type as an input type", "GRAPHQL_114",
-                                            DiagnosticSeverity.ERROR),
-        INVALID_LOCATION_FOR_CONTEXT_PARAMETER("The graphql:Context should be the first parameter", "GRAPHQL_115",
-                                               DiagnosticSeverity.ERROR);
-
-        private final String error;
-        private final String errorCode;
-        private final DiagnosticSeverity diagnosticSeverity;
-
-        CompilationError(String error, String errorCode, DiagnosticSeverity diagnosticSeverity) {
-            this.error = error;
-            this.errorCode = errorCode;
-            this.diagnosticSeverity = diagnosticSeverity;
-        }
-
-        String getError() {
-            return error;
-        }
-
-        String getErrorCode() {
-            return errorCode;
-        }
-
-        DiagnosticSeverity getDiagnosticSeverity() {
-            return this.diagnosticSeverity;
-        }
-    }
+    public static final String FILE_UPLOAD_IDENTIFIER = "Upload";
 
     public static boolean validateModuleId(ModuleSymbol moduleSymbol) {
         String moduleName = moduleSymbol.id().moduleName();
@@ -139,14 +72,6 @@ public class Utils {
         return methodSymbol.qualifiers().contains(Qualifier.REMOTE);
     }
 
-    public static void updateContext(SyntaxNodeAnalysisContext context, CompilationError errorCode,
-                                     Location location) {
-        DiagnosticInfo diagnosticInfo = new DiagnosticInfo(
-                errorCode.getErrorCode(), errorCode.getError(), errorCode.getDiagnosticSeverity());
-        Diagnostic diagnostic = DiagnosticFactory.createDiagnostic(diagnosticInfo, location);
-        context.reportDiagnostic(diagnostic);
-    }
-
     public static boolean isGraphqlListener(TypeSymbol listenerType) {
         if (listenerType.typeKind() == TypeDescKind.UNION) {
             return ((UnionTypeSymbol) listenerType).memberTypeDescriptors().stream()
@@ -168,16 +93,5 @@ public class Utils {
             return moduleOpt.isPresent() && validateModuleId(moduleOpt.get());
         }
         return false;
-    }
-
-    public static boolean isInvalidFieldName(String fieldName) {
-        return fieldName.startsWith(DOUBLE_UNDERSCORES) || fieldName.contains("/" + DOUBLE_UNDERSCORES);
-    }
-
-    static Location getLocation(Symbol typeSymbol, Location alternateLocation) {
-        if (typeSymbol.getLocation().isPresent()) {
-            return typeSymbol.getLocation().get();
-        }
-        return alternateLocation;
     }
 }